refactor to lay the groundwork and structure for #64, #59 and #58
