+++ src/main/scala/org/cakesolutions/scalad/experimental/MongoNativeRestrictions.scala
@@ -1,8 +1,9 @@
-package org.cakesolutions.scalad.mongo
+package org.cakesolutions.scalad.experimental
 
-import spray.json.{JsString, JsObject}
 import org.cakesolutions.scalad._
 import com.mongodb.{BasicDBObjectBuilder, DBObject}
+import mongo.MongoSerializer
+import mongo.sprayjson.SprayJsonSerializers
 
 /**
  * Native restrictions for JsObject and MongoDB
@@ -13,10 +14,10 @@ trait MongoJsonNativeRestrictions extends NativeRestrictions with MongoNativeRes
   def convertToNative(restriction: Restriction) = {
     def convert0(builder: BasicDBObjectBuilder, r: Restriction) {
       r match {
-        case EqualsRestriction(path: String, value)    => builder.add(path, value)
+        case EqualsRestriction(path: String, value) => builder.add(path, value)
         case NotEqualsRestriction(path: String, value) => builder.add(path, value)
-        case ConjunctionRestriction(lhs, rhs)          => convert0(builder, lhs); convert0(builder, rhs)
-        case DisjunctionRestriction(lhs, rhs)          => convert0(builder, lhs); convert0(builder, rhs)
+        case ConjunctionRestriction(lhs, rhs) => convert0(builder, lhs); convert0(builder, rhs)
+        case DisjunctionRestriction(lhs, rhs) => convert0(builder, lhs); convert0(builder, rhs)
         // ...
       }
     }
@@ -28,9 +29,9 @@ trait MongoJsonNativeRestrictions extends NativeRestrictions with MongoNativeRes
 
 }
 
-private[mongo] trait MongoNativeRestrictionMarshallers extends SprayJsonSerializers {
+private[experimental] trait MongoNativeRestrictionMarshallers extends SprayJsonSerializers {
 
-  implicit def getNativeRestrictionsMarshaller[A : MongoSerializer]: NativeRestrictionsMarshaller[A] = new NativeRestrictionsMarshaller[A] {
+  implicit def getNativeRestrictionsMarshaller[A: MongoSerializer]: NativeRestrictionsMarshaller[A] = new NativeRestrictionsMarshaller[A] {
     val serializer = implicitly[MongoSerializer[A]]
 
     type NativeRestrictionValue = DBObject
