+++ src/main/scala/org/cakesolutions/scalad/mongo/sprayjson/crud.scala
@@ -1,0 +1,2 @@
+package org.cakesolutions.scalad.mongo.sprayjson
+
